pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.37.0",
    "oauthlib>=3.3.1",
    "pypdf2>=3.0.1",
    "sqlalchemy>=2.0.43",
    "pdfplumber>=0.11.0",
    "spacy>=3.7.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]
